tkm-sr-pro/backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
django==4.2.5
djangorestframework==3.14.0
psycopg2-binary==2.9.7
elasticsearch==8.10.0
transformers==4.33.1
spacy==3.6.1
sentence-transformers==2.2.2
pdfplumber==0.10.2
python-Levenshtein==0.21.1
scikit-learn==1.3.0
numpy<2.0.0
django-cors-headers==4.2.0
python-dotenv==1.0.0
